+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Fix: Merge user from event and scope (#467)
+
 # 5.1.0-beta.1
 
 * Fix: `Sentry.close()` closes native SDK integrations (#388)